+++ bonus/src/controls/key_actions/player_shoot.c
@@ -1,23 +1,24 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   hooks.c                                            :+:      :+:    :+:   */
+/*   player_shoot.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: sfrankie <sfrankie@student.42wolfsburg.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/05/29 12:11:37 by sfrankie          #+#    #+#             */
-/*   Updated: 2024/07/31 12:54:19 by sfrankie         ###   ########.fr       */
+/*   Created: 2024/08/05 00:40:22 by sfrankie          #+#    #+#             */
+/*   Updated: 2024/08/05 00:40:23 by sfrankie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../incl/cub3d.h"
+#include "../../../incl/cub3d.h"
 
-void	init_hooks(t_game *game)
+void	action_mouse_left_click(t_game *game)
 {
-	mlx_hook(game->win_ptr, 02, 1L << 0, keypress, game);
-	mlx_hook(game->win_ptr, 3, 1L << 1, keyrelease, game);
-	mlx_hook(game->win_ptr, 6, 1L << 6, mouse_move, game);
-	mlx_hook(game->win_ptr, 17, 0, close_game, game);
-	mlx_mouse_hook(game->win_ptr, mouse_press, game);
-	mlx_loop_hook(game->mlx_ptr, loop_hook, game);
+	game->animation.animation_gun = true;
+	if (game->player == PABECKHA)
+		animation_gun_shoot(game, SHOTGUN_FRAME_DURATION, SHOTGUN_FRAMES,
+							game->textures.shotgun);
+	else
+		animation_gun_shoot(game, DESERT_EAGLE_FRAME_DURATION,
+							DESERT_EAGLE_FRAMES, game->textures.desert_eagle);
 }
