+++ bonus/src/ft_free.c
@@ -12,19 +12,19 @@
 
 #include "../incl/cub3d.h"
 
-void free_2d_array(char **array)
+void	free_2d_array(char **array)
 {
-    int i;
+	int	i;
 
-    i = 0;
-    if (array)
-    {
-        while(array[i])
-        {
-            free(array[i]);
-            i++;
-        }
-        free(array);
-    }
-    array = NULL;
+	i = 0;
+	if (array)
+	{
+		while (array[i])
+		{
+			free(array[i]);
+			i++;
+		}
+		free(array);
+	}
+	array = NULL;
 }
