+++ bonus/src/init_game.c
@@ -6,13 +6,78 @@
 /*   By: sfrankie <sfrankie@student.42wolfsburg.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/05 12:25:26 by sfrankie          #+#    #+#             */
-/*   Updated: 2024/07/31 12:55:21 by sfrankie         ###   ########.fr       */
+/*   Updated: 2024/08/05 01:22:03 by sfrankie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../incl/cub3d.h"
 
 void	init_game(t_game *game)
+{
+	int	i;
+
+	init_mlx(game);
+	init_image(game);
+	init_utils(game);
+	game->animation.door_are_opening = false;
+	game->animation.door_are_closing = false;
+	game->animation.hp_frame_updated = false;
+	game->animation.animation_gun = false;
+	i = 0;
+	while (i < NUM_KEYS)
+		game->keys[i++] = false;
+	i = 0;
+	while (i < DEPTH_MAX)
+	{
+		game->door[i].y = 0;
+		game->door[i].x = 0;
+		game->enemy[i].x = 0;
+		game->enemy[i].y = 0;
+		game->enemy[i].dead = false;
+		++i;
+	}
+	calc_dir_vectors(game);
+}
+
+void	init_utils(t_game *game)
+{
+	if (game->data.player->direction == 'N')
+		game->ray_main_angle = 0.75f * ANGLE_MAX;
+	else if (game->data.player->direction == 'W')
+		game->ray_main_angle = 0.5f * ANGLE_MAX;
+	else if (game->data.player->direction == 'S')
+		game->ray_main_angle = 0.25f * ANGLE_MAX;
+	else
+		game->ray_main_angle = 0;
+	game->ray_hit_x = 0;
+	game->ray_hit_y = 0;
+	game->ray_new_x = 0;
+	game->ray_new_y = 0;
+	game->dist_idx = 0;
+	game->wall_direction = 0;
+	game->door_direction = 0;
+	game->img_x = 0;
+	game->img_y = 0;
+	game->mouse_x = 0;
+	game->player_dead = false;
+	game->main_menu = true;
+	game->enemy_count = 0;
+	game->won_game = false;
+	game->player = PABECKHA;
+}
+
+void	init_image(t_game *game)
+{
+	game->image.img = mlx_new_image(game->mlx_ptr, SCREEN_WIDTH, SCREEN_HEIGHT);
+	game->image.bits_per_pixel = 32;
+	game->image.line_length = SCREEN_WIDTH;
+	game->image.endian = 0;
+	game->image.data = (int *)mlx_get_data_addr(game->image.img,
+			&game->image.bits_per_pixel, &game->image.line_length,
+			&game->image.endian);
+}
+
+void	init_mlx(t_game *game)
 {
 	game->mlx_ptr = mlx_init();
 	if (!game->mlx_ptr)
@@ -27,66 +92,28 @@ void	init_game(t_game *game)
 		perror("mlx window");
 		exit(EXIT_FAILURE);
 	}
-	game->ray_new_x = 0;
-	game->ray_new_y = 0;
-	if (game->data->player->direction == 'N')
-		game->ray_main_angle = 0.75f * ANGLE_MAX;
-	else if (game->data->player->direction == 'W')
-		game->ray_main_angle = 0.5f * ANGLE_MAX;
-	else if (game->data->player->direction == 'S')
-		game->ray_main_angle = 0.25f * ANGLE_MAX;
-	else
-		game->ray_main_angle = 0;
-	game->ray_hit_x = 0;
-	game->ray_hit_y = 0;
-	game->dist_idx = 0;
-	game->wall_direction = 0;
-	game->door_direction = 0;
-	game->img_x = 0;
-	game->img_y = 0;
-	init_keys(game);
-	game->door_are_opening = false;
-	game->door_are_closing = false;
-	game->mouse_x = 0;
-	game->hp_frame_updated = false;
-	game->player_dead = false;
-	game->depth_lvl = 0;
-	game->dist_idx = 0;
-	while (game->depth_lvl < DEPTH_MAX)
-	{
-		game->depth[game->depth_lvl].dist = 0;
-		game->depth[game->depth_lvl].obj = EMPTY;
-		game->depth[game->depth_lvl].ray_hit_y = 0;
-		game->depth[game->depth_lvl].ray_hit_x = 0;
-		game->depth[game->depth_lvl].colis_y = 0;
-		game->depth[game->depth_lvl].colis_x = 0;
-		game->door[game->depth_lvl].y = 0;
-		game->door[game->depth_lvl].x = 0;
-		game->enemy[game->depth_lvl].x = 0;
-		game->enemy[game->depth_lvl].y = 0;
-		game->enemy[game->depth_lvl].x_start = 0;
-		game->enemy[game->depth_lvl].x_end = 0;
-		game->enemy[game->depth_lvl].size = 0;
-		game->enemy[game->depth_lvl].tex_x = 0;
-		game->enemy[game->depth_lvl].visible = false;
-		game->enemy[game->depth_lvl].dead = false;
-		++game->depth_lvl;
-	}
-	game->depth_lvl = 0;
-	game->dist_idx = 0;
-	game->main_menu = true;
-	game->animation_gun = false;
-	game->enemy_count = 0;
-	game->won_game = false;
-	game->player = PABECKHA;
-	game->enemy_visible = false;
 }
 
-void	init_keys(t_game *game)
+void	calc_dir_vectors(t_game *game)
 {
-	int i;
+	float	angle_incr_radians;
+	float	angle_iter;
 
-	i = 0;
-	while (i < NUM_KEYS)
-		game->keys[i++] = false;
+	angle_incr_radians = to_radians(ROTATION_SPEED);
+	angle_iter = to_radians(game->ray_main_angle);
+	if (angle_iter < 0)
+		angle_iter += 2 * M_PI;
+	else if (angle_iter > 2 * M_PI)
+		angle_iter -= 2 * M_PI;
+	game->vec_idx = 0;
+	while (game->vec_idx < ANGLE_MAX / ROTATION_SPEED)
+	{
+		game->vectors[game->vec_idx].x = cosf(angle_iter);
+		game->vectors[game->vec_idx].y = sinf(angle_iter);
+		game->vec_idx++;
+		angle_iter += angle_incr_radians;
+		if (angle_iter >= 2 * M_PI)
+			angle_iter -= 2 * M_PI;
+	}
+	game->vec_idx = 0;
 }
