+++ bonus/src/parser/map_utils.c
@@ -73,7 +73,6 @@ bool	check_map_element_input(t_data *data)
 	int	i;
 	int	j;
 
-	j = 0;
 	i = data->line_start_map_position + 1;
 	while (data->cub_file[i])
 	{
@@ -81,13 +80,12 @@ bool	check_map_element_input(t_data *data)
 		while (data->cub_file[i][j])
 		{
 			if (data->cub_file[i][j] != '0' && data->cub_file[i][j] != '1'
-			    && data->cub_file[i][j] != ' ' && data->cub_file[i][j] != 'N'
-			    && data->cub_file[i][j] != 'S' && data->cub_file[i][j] != 'E'
-			    && data->cub_file[i][j] != 'W' && data->cub_file[i][j] != '2'
-			    && data->cub_file[i][j] != '3' && data->cub_file[i][j] != '4'
-			    && data->cub_file[i][j] != '6'
-			    && (data->cub_file[i][j] < 9
-			        || data->cub_file[i][j] > 13))
+				&& data->cub_file[i][j] != ' ' && data->cub_file[i][j] != 'N'
+				&& data->cub_file[i][j] != 'S' && data->cub_file[i][j] != 'E'
+				&& data->cub_file[i][j] != 'W' && data->cub_file[i][j] != '2'
+				&& data->cub_file[i][j] != '3' && data->cub_file[i][j] != '4'
+				&& data->cub_file[i][j] != '6' && (data->cub_file[i][j] < 9
+				|| data->cub_file[i][j] > 13))
 			{
 				replace_error_message(data, "Map element in wrong format");
 				return (false);
